Reorganize handlers across API categories
Now that Ook is starting to add fairly diverse types of APIs, it makes
sense to start splitting the APIs into different modules. This new
orgnanization provides dedicated routers for /link and /ingest
endpoints. It also provides each API category with its own models module
to maintain the API models related to those endpoints.
