--- src/ook/handlers/external/__init__.py
@@ -1,3 +1,0 @@
-"""Externally-accessible endpoint handlers that serve relative to
-``/<app-name>/``.
-"""
