+++ src/ook/handlers/ingest/__init__.py
@@ -1,0 +1,7 @@
+"""Ingest API."""
+
+from .endpoints import router as ingest_router
+
+__all__ = [
+    "ingest_router",
+]
