+++ src/ook/handlers/internal/__init__.py
@@ -1,6 +1,9 @@
 """Internal HTTP handlers that serve relative to the root path, ``/``.
 
 These handlers aren't externally visible since the app is available at a path,
-``/ook``. See `ook.handlers.external` for
-the external endpoint handlers.
+``/ook``.
 """
+
+from .endpoints import router as internal_router
+
+__all__ = ["internal_router"]
