+++ src/ook/handlers/links/__init__.py
@@ -1,0 +1,7 @@
+"""Links API."""
+
+from .endpoints import router as links_router
+
+__all__ = [
+    "links_router",
+]
