+++ tests/handlers/internal/internal_endpoints_test.py
